[#45 #27] Configurado docker con bbdd, pasa tests de Postman
